webapp/requirements.txt
flask
langchain-groq
langchain-core
google-api-python-client
google-cloud-bigquery
google-cloud-storage
numpy
pandas
tabulate
db-dtypes